Revert "Refactor: Made linedef_s an opaque type, moved members to LineDef class"
This reverts commit 1b882f44e5afcbb2d91c2fc3def06c253db03b83.

+++ doomsday/engine/include/gl/gl_main.h
@@ -58,7 +58,7 @@ extern "C" {
     GLint p; \
     glGetIntegerv(GL_TEXTURE_BINDING_2D, &p); \
     Sys_GLCheckError(); \
-    assert(p == (GLint)tex); \
+    assert(p == tex); \
 }
 #else
 #  define LIBDENG_ASSERT_GL_TEXTURE_ISBOUND(tex)
